+++ ravenpy/models/__init__.py
@@ -3,7 +3,7 @@ import os
 from .base import Ostrich, Raven, get_average_annual_runoff
 from .emulators import *
 from .multimodel import RavenMultiModel
-from .rv import HRU, LU, RV, RVI, HRUState, Sub
+from .rv import HRU, LU, RV, HRUState, Sub
 
 _dir = os.path.abspath(os.path.dirname(__file__))
