--- ravenpy/models/raven-gr4j-cemaneige/raven-gr4j-cemaneige.rvc
@@ -1,1 +1,0 @@
-../global/global.rvc
