--- ravenpy/models/raven-gr4j-cemaneige/raven-gr4j-cemaneige.rvh
@@ -1,22 +1,0 @@
-#########################################################################
-:FileType          rvh ASCII Raven 2.8.2
-:WrittenBy         Juliane Mai & James Craig
-:CreationDate      Sep 2018
-#
-# Emulation of GR4J simulation of Salmon River near Prince George
-#------------------------------------------------------------------------
-#
-
-{subbasins_cmd}
-
-{hrus_cmd}
-
-{land_subbasin_group_cmd}
-
-{land_subbasin_property_multiplier}
-
-{lake_subbasin_group_cmd}
-
-{lake_subbasin_property_multiplier}
-
-{reservoir_cmd_list}
