--- ravenpy/models/raven-gr4j-cemaneige/raven-gr4j-cemaneige.rvi
@@ -1,76 +1,0 @@
-#########################################################################
-:FileType          rvi ASCII Raven 2.8.2
-:WrittenBy         Juliane Mai & James Craig
-:CreationDate      Sep 2018
-#
-# Emulation of GR4J simulation of Salmon River near Prince George
-#------------------------------------------------------------------------
-#
-
-:Calendar              {calendar}
-:RunName               {run_name}-{run_index}
-:StartDate             {start_date}
-:EndDate               {end_date}
-:TimeStep              {time_step}
-:Method                ORDERED_SERIES
-
-:SoilModel             SOIL_MULTILAYER  4
-{routing_cmd}
-:CatchmentRoute        ROUTE_DUMP
-:Evaporation           {evaporation}  # PET_OUDIN
-:RainSnowFraction      {rain_snow_fraction}  # RAINSNOW_DINGMAN
-:PotentialMeltMethod   POTMELT_DEGREE_DAY
-:OroTempCorrect        OROCORR_SIMPLELAPSE
-:OroPrecipCorrect      OROCORR_SIMPLELAPSE
-
-#------------------------------------------------------------------------
-# Soil Layer Alias Definitions
-#
-:Alias PRODUCT_STORE      SOIL[0]
-:Alias ROUTING_STORE      SOIL[1]
-:Alias TEMP_STORE         SOIL[2]
-:Alias GW_STORE           SOIL[3]
-
-#------------------------------------------------------------------------
-# Hydrologic process order for GR4J Emulation
-#
-:HydrologicProcesses
- :Precipitation            PRECIP_RAVEN       ATMOS_PRECIP    MULTIPLE
- :SnowTempEvolve           SNOTEMP_NEWTONS    SNOW_TEMP
- :SnowBalance              SNOBAL_CEMA_NIEGE  SNOW            PONDED_WATER
- :OpenWaterEvaporation     OPEN_WATER_EVAP    PONDED_WATER    ATMOSPHERE     			 # Pn
- :Infiltration             INF_GR4J           PONDED_WATER    MULTIPLE       			 # Ps-
- :SoilEvaporation          SOILEVAP_GR4J      PRODUCT_STORE   ATMOSPHERE     			 # Es
- :Percolation              PERC_GR4J          PRODUCT_STORE   TEMP_STORE     			 # Perc
- :Flush                    RAVEN_DEFAULT      SURFACE_WATER   TEMP_STORE     			 # Pn-Ps
- :Split                    RAVEN_DEFAULT      TEMP_STORE      CONVOLUTION[0] CONVOLUTION[1] 0.9  # Split Pr
- :Convolve                 CONVOL_GR4J_1      CONVOLUTION[0]  ROUTING_STORE  			 # Q9
- :Convolve                 CONVOL_GR4J_2      CONVOLUTION[1]  TEMP_STORE     			 # Q1
- :Percolation              PERC_GR4JEXCH      ROUTING_STORE   GW_STORE       			 # F(x1)
- :Percolation              PERC_GR4JEXCH2     TEMP_STORE      GW_STORE       			 # F(x1)
- :Flush                    RAVEN_DEFAULT      TEMP_STORE      SURFACE_WATER  			 # Qd
- :Baseflow                 BASE_GR4J          ROUTING_STORE   SURFACE_WATER  			 # Qr
-:EndHydrologicProcesses
-#------------------------------------------------------------------------
-
-#---------------------------------------------------------
-# Output Options
-#
-:WriteForcingFunctions
-:EvaluationMetrics {evaluation_metrics}
-:WriteNetcdfFormat  yes
-#:NoisyMode
-:SilentMode
-:PavicsMode
-{suppress_output}
-
-:NetCDFAttribute title Simulated river discharge
-:NetCDFAttribute history Created on {now} by Raven
-:NetCDFAttribute references  Craig, J.R., and the Raven Development Team, Raven user's and developer's manual (Version 2.8), URL: http://raven.uwaterloo.ca/ (2018).
-:NetCDFAttribute comment Raven Hydrological Framework version {raven_version}
-
-:NetCDFAttribute model_id gr4jcn
-
-:NetCDFAttribute time_frequency day
-:NetCDFAttribute time_coverage_start {start_date}
-:NetCDFAttribute time_coverage_end {end_date}
