--- ravenpy/models/raven-gr4j-cemaneige/raven-gr4j-cemaneige.rvt
@@ -1,1 +1,0 @@
-../global/global.rvt
