--- ravenpy/models/raven-routing/raven-routing.rvi
@@ -1,45 +1,0 @@
-# ----------------------------------------------
-# Raven Input file
-# Grid-based routing example
-# ----------------------------------------------
-
-:Calendar              {calendar}
-:RunName               {run_name}-{run_index}
-:StartDate             {start_date}
-:EndDate               {end_date}
-:TimeStep              {time_step}
-:Method                ORDERED_SERIES                # Numerical method used for simulation
-
-:CatchmentRoute        ROUTE_DUMP                    # Catchment routing method, used to convey water from the catchment tributaries and rivulets to the subbasin outlets. DEFAULT ROUTE_DUMP, which instantly ‘dumps’ all water in the subbasin stream reach.
-:Routing               ROUTE_DIFFUSIVE_WAVE          # Channel routing method which is used to transport water from upstream to downstream within the main subbasin channels. DEFAULT ROUTE_DIFFUSIVE_WAVE, which analytically solves the diffusive wave equation along the reach using a constant reference celerity.
-:PrecipIceptFract      PRECIP_ICEPT_NONE             # Estimation of the precipitation interception fraction. In this routing model, stream input(s) are "pretending" to be precipitation going into Raven, thus using DEFAULT PRECIP_ICEPT_NONE to indicate no interception processes are adopted.
-:PotentialMeltMethod   POTMELT_NONE                  # Estimation of the potential snow melt. In this routing model, snow melt processes are not relevant, thus using DEFAULT POTMELT_NONE method.
-:SoilModel             SOIL_ONE_LAYER                # In this routing model, use DEFAULT SOIL_ONE_LAYER to define single soil layer structure.
-
-:HydrologicProcesses
-  :Precipitation     PRECIP_RAVEN             ATMOS_PRECIP     PONDED_WATER          # Moves stream input(s) from ATMOS_PRECIP to PONDED_WATER storage (waiting for runoff). Use DEFAULT PRECIP_RAVEN method.
-  :Flush             RAVEN_DEFAULT            PONDED_WATER     SURFACE_WATER         # Moves water from PONDED_WATER to SURFACE_WATER (routed to outlet). Use DEFAULT RAVEN_DEFAULT method.
-:EndHydrologicProcesses
-
-
-# Output Options
-#
-#:WriteForcingFunctions
-# Defines the hydrograph performance metrics output by Raven. Either one or multiple is acceptable.
-:EvaluationMetrics {evaluation_metrics}
-:WriteNetcdfFormat  yes
-#:NoisyMode
-:SilentMode
-:PavicsMode
-{suppress_output}
-
-:NetCDFAttribute title Simulated river discharge
-:NetCDFAttribute history Created on {now} by Raven
-:NetCDFAttribute references  Craig, J.R., and the Raven Development Team, Raven user's and developer's manual (Version 2.8), URL: http://raven.uwaterloo.ca/ (2018).
-:NetCDFAttribute comment Raven Hydrological Framework version {raven_version}
-
-:NetCDFAttribute model_id routing
-
-:NetCDFAttribute time_frequency day
-:NetCDFAttribute time_coverage_start {start_date}
-:NetCDFAttribute time_coverage_end {end_date}
