--- ravenpy/models/raven-routing/raven-routing.rvt
@@ -1,8 +1,0 @@
-# ----------------------------------------------
-# Raven Input file
-# Walkthrough #6: Grid-based routing example
-# ----------------------------------------------
-
-{gridded_forcing_list}
-
-{observation_data}
