+++ ravenpy/models/rv.py
@@ -63,84 +63,6 @@ Simulation end date and duration are updated automatically when duration, start 
 
 """
 
-# CF standard names used for mappings
-default_input_variables = (
-    "pr",
-    "rainfall",
-    "prsn",
-    "tasmin",
-    "tasmax",
-    "tas",
-    "evspsbl",
-    "water_volume_transport_in_river_channel",
-)
-
-# # Map CF-Convention standard name to Raven Forcing name
-# forcing_names = {
-#     "tasmin": "TEMP_MIN",
-#     "tasmax": "TEMP_MAX",
-#     "tas": "TEMP_AVE",
-#     "rainfall": "RAINFALL",
-#     "pr": "PRECIP",
-#     "prsn": "SNOWFALL",
-#     "evspsbl": "PET",
-#     "water_volume_transport_in_river_channel": "HYDROGRAPH",
-# }
-
-# # Alternate typical variable names found in netCDF files, keyed by CF standard name
-# alternate_nc_names = {
-#     "tasmin": ["tasmin", "tmin"],
-#     "tasmax": ["tasmax", "tmax"],
-#     "tas": ["tas", "t2m"],
-#     "rainfall": ["rainfall", "rain"],
-#     "pr": ["pr", "precip", "prec", "precipitation", "tp"],
-#     "prsn": ["prsn", "snow", "snowfall", "solid_precip"],
-#     "evspsbl": ["pet", "evap", "evapotranspiration"],
-#     "water_volume_transport_in_river_channel": [
-#         "qobs",
-#         "discharge",
-#         "streamflow",
-#         "dis",
-#     ],
-# }
-
-rain_snow_fraction_options = (
-    "RAINSNOW_DATA",
-    "RAINSNOW_DINGMAN",
-    "RAINSNOW_UBC",
-    "RAINSNOW_HBV",
-    "RAINSNOW_HARDER",
-    "RAINSNOW_HSPF",
-)
-
-evaporation_options = (
-    "PET_CONSTANT",
-    "PET_PENMAN_MONTEITH",
-    "PET_PENMAN_COMBINATION",
-    "PET_PRIESTLEY_TAYLOR",
-    "PET_HARGREAVES",
-    "PET_HARGREAVES_1985",
-    "PET_FROMMONTHLY",
-    "PET_DATA",
-    "PET_HAMON_1961",
-    "PET_TURC_1961",
-    "PET_MAKKINK_1957",
-    "PET_MONTHLY_FACTOR",
-    "PET_MOHYSE",
-    "PET_OUDIN",
-)
-
-calendar_options = (
-    "PROLEPTIC_GREGORIAN",
-    "JULIAN",
-    "GREGORIAN",
-    "STANDARD",
-    "NOLEAP",
-    "365_DAY",
-    "ALL_LEAP",
-    "366_DAY",
-)
-
 
 class RVFile:
     def __init__(self, fn):
@@ -266,219 +188,6 @@ class RV(collections.abc.Mapping, RavenConfig):
         else:
             for key, val in items.items():
                 self[key] = val
-
-
-class RVI(RV):
-    def __init__(self, **kwargs):
-        self.name = None
-        self.area = None
-        self.elevation = None
-        self.latitude = None
-        self.longitude = None
-        self.run_index = 0
-        self.raven_version = "3.0.1 rev#275"
-        self.routing = "ROUTE_NONE"
-
-        self._run_name = "run"
-        self._start_date = None
-        self._end_date = None
-        self._now = None
-        self._rain_snow_fraction = "RAINSNOW_DATA"
-        self._evaporation = None
-        self._ow_evaporation = None
-        self._duration = 1
-        self._time_step = 1.0
-        self._evaluation_metrics = "NASH_SUTCLIFFE RMSE"
-        self._suppress_output = False
-        self._calendar = "standard"
-
-        super(RVI, self).__init__(**kwargs)
-
-    @property
-    def run_name(self):
-        return self._run_name
-
-    @run_name.setter
-    def run_name(self, x):
-        if isinstance(x, six.string_types):
-            self._run_name = x
-        else:
-            raise ValueError("Must be string")
-
-    @property
-    def start_date(self):
-        return self._start_date
-
-    @start_date.setter
-    def start_date(self, x):
-        if isinstance(x, dt.datetime):
-            self._start_date = self._dt2cf(x)
-        else:
-            raise ValueError("Must be datetime")
-
-        if x == dt.datetime(1, 1, 1):
-            return
-
-        if self._duration is None:
-            self._update_duration()
-        else:
-            self._update_end_date()
-
-    @property
-    def end_date(self):
-        return self._end_date
-
-    @end_date.setter
-    def end_date(self, x):
-        if isinstance(x, dt.datetime):
-            self._end_date = self._dt2cf(x)
-        else:
-            raise ValueError("Must be datetime")
-
-        if x != dt.datetime(1, 1, 1):
-            self._update_duration()
-
-    @property
-    def duration(self):
-        return self._duration
-
-    @duration.setter
-    def duration(self, x):
-        if isinstance(x, int):
-            if x > 0:
-                self._duration = x
-        else:
-            raise ValueError("Must be int")
-
-        if x > 0:
-            self._update_end_date()
-
-    @property
-    def time_step(self):
-        return self._time_step
-
-    @time_step.setter
-    def time_step(self, x):
-        self._time_step = x
-
-    @property
-    def evaluation_metrics(self):
-        return self._evaluation_metrics
-
-    @evaluation_metrics.setter
-    def evaluation_metrics(self, x):
-        if not isinstance(x, six.string_types):
-            raise ValueError("Evaluation metrics must be string.")
-
-        for metric in x.split():
-            if metric not in {
-                "NASH_SUTCLIFFE",
-                "LOG_NASH",
-                "RMSE",
-                "PCT_BIAS",
-                "ABSERR",
-                "ABSMAX",
-                "PDIFF",
-                "TMVOL",
-                "RCOEFF",
-                "NSC",
-                "KLING_GUPTA",
-            }:
-                raise ValueError("{} is not a metric recognized by Raven.")
-
-        self._evaluation_metrics = x
-
-    def _update_duration(self):
-        if self.end_date is not None and self.start_date is not None:
-            self._duration = (self.end_date - self.start_date).days
-
-    def _update_end_date(self):
-        if self.start_date is not None and self.duration is not None:
-            self._end_date = self.start_date + dt.timedelta(days=self.duration)
-
-    @property
-    def now(self):
-        return dt.datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-
-    @property
-    def suppress_output(self):
-        tag = ":SuppressOutput\n:DontWriteWatershedStorage"
-        return tag if self._suppress_output else ""
-
-    @suppress_output.setter
-    def suppress_output(self, value):
-        if not isinstance(value, bool):
-            raise ValueError
-        self._suppress_output = value
-
-    @property
-    def routing_cmd(self):
-        return RoutingCommand(value=self.routing)
-
-    @property
-    def rain_snow_fraction(self):
-        """Rain snow partitioning."""
-        return self._rain_snow_fraction
-
-    @rain_snow_fraction.setter
-    def rain_snow_fraction(self, value):
-        """Can be one of
-
-        - RAINSNOW_DATA
-        - RAINSNOW_DINGMAN
-        - RAINSNOW_UBC
-        - RAINSNOW_HBV
-        - RAINSNOW_HARDER
-        - RAINSNOW_HSPF
-        """
-        v = value.upper()
-
-        if v in rain_snow_fraction_options:
-            self._rain_snow_fraction = v
-        else:
-            raise ValueError(f"Value should be one of {rain_snow_fraction_options}.")
-
-    @property
-    def evaporation(self):
-        """Evaporation scheme"""
-        return self._evaporation
-
-    @evaporation.setter
-    def evaporation(self, value):
-        v = value.upper()
-        if v in evaporation_options:
-            self._evaporation = v
-        else:
-            raise ValueError(f"Value {v} should be one of {evaporation_options}.")
-
-    @property
-    def ow_evaporation(self):
-        """Open-water evaporation scheme"""
-        return self._ow_evaporation
-
-    @ow_evaporation.setter
-    def ow_evaporation(self, value):
-        v = value.upper()
-        if v in evaporation_options:
-            self._ow_evaporation = v
-        else:
-            raise ValueError(f"Value {v} should be one of {evaporation_options}.")
-
-    @property
-    def calendar(self):
-        """Calendar"""
-        return self._calendar.upper()
-
-    @calendar.setter
-    def calendar(self, value):
-        if value.upper() in calendar_options:
-            self._calendar = value
-        else:
-            raise ValueError(f"Value should be one of {calendar_options}.")
-
-    def _dt2cf(self, date):
-        """Convert datetime to cftime datetime."""
-        return cftime._cftime.DATE_TYPES[self._calendar.lower()](*date.timetuple()[:6])
 
 
 class Ost(RV):
